[FIX] fix Swath window estimation
fix how OpenSWATH estimates the number of Swath windows present: instead of
assuming that an MS1 scan is only acquired after a full cycle, now the cycle of
Swath windows is defined as returning to the original window. This completes
the cycle and allows for more accurate determination of Swath window setup in
different Swath acquisition setups.